New coordinate route line/column + (WIP) specific data set

+++ src-api/server/database/migrations/2020_01_06_130000_create_coordinates_table.php
@@ -17,8 +17,8 @@ class CreateCoordinatesTable extends Migration
             $table->bigIncrements('id')->unsigned();
             $table->float('latitude')->unsigned();
             $table->float('longitude')->unsigned();
-            $table->float('line')->unsigned();
-            $table->float('column')->unsigned();
+            $table->integer('line')->unsigned();
+            $table->integer('column')->unsigned();
             $table->timestamps();
         });
     }
